Proposal for an expedited deprecation process.
For a long time now, the formal deprecation process for GTNH has been "notify players that this is being deprecated, for one stable release, and terminate it the next dev cycle. This directly means that it usually takes up to a year to deprecate what is often fully unfunctional content such as Bart's original LEG from EMT.

Likewise, there has been a few moments where this process as a whole has been completely ignored in favour of an instant termination of service for a specific bit of content. One recent example of this was the PA debacle from earlier this cycle (that was later reversed).

There are many moments in which an expedited process of termination should be in effect, for example, the aforementioned EMT LEG which took nearly a year and a half after being replaced just to terminate.

As such, I would like to set in stone some guidelines to enact a formalized process for instant termination within certain cases, such as Bart's LEG or Hydropower's unfunctional and instance corrupting dam multi.